レビュー修正

+++ CRUD/ViewController/BookListCell.swift
@@ -14,21 +14,21 @@ class BookListCell: UITableViewCell {
     
     lazy var title: UILabel = {
         let title = UILabel()
-        title.font = .systemFont(ofSize: CGFloat(NumberManager.fontSizeSmall))
+        title.font = .systemFont(ofSize: NumberManager.fontSizeSmall)
         title.translatesAutoresizingMaskIntoConstraints = false
         return title
     }()
     
     lazy var price: UILabel = {
         let price = UILabel()
-        price.font = .systemFont(ofSize: CGFloat(NumberManager.fontSizeSmall))
+        price.font = .systemFont(ofSize: NumberManager.fontSizeSmall)
         price.translatesAutoresizingMaskIntoConstraints = false
         return price
     }()
     
     lazy var bookDay: UILabel = {
         let bookDay = UILabel()
-        bookDay.font = .systemFont(ofSize: CGFloat(NumberManager.fontSizeSmall))
+        bookDay.font = .systemFont(ofSize: NumberManager.fontSizeSmall)
         bookDay.translatesAutoresizingMaskIntoConstraints = false
         return bookDay
     }()
@@ -37,7 +37,7 @@ class BookListCell: UITableViewCell {
         let button = UIButton()
         button.setTitle(">", for: .normal)
         button.setTitleColor(.black, for: .normal)
-        button.titleLabel?.font = .systemFont(ofSize: CGFloat(NumberManager.fontSizeLarge))
+        button.titleLabel?.font = .systemFont(ofSize: NumberManager.fontSizeLarge)
         button.translatesAutoresizingMaskIntoConstraints = false
         return button
     }()
@@ -55,25 +55,25 @@ class BookListCell: UITableViewCell {
 extension BookListCell {
     private func setupUI() {
         contentView.addSubview(bookimage)
-        bookimage.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: CGFloat(NumberManager.bookImageTopConstraint)).isActive = true
-        bookimage.leftAnchor.constraint(equalTo: self.contentView.leftAnchor, constant: CGFloat(NumberManager.bookImageLeftConstraintInCell)).isActive = true
-        bookimage.widthAnchor.constraint(equalToConstant: CGFloat(NumberManager.bookImageWidthConstraintInCell)).isActive = true
-        bookimage.heightAnchor.constraint(equalToConstant: CGFloat(NumberManager.bookImageHeightConstraintInCell)).isActive = true
+        bookimage.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: NumberManager.bookImageTopConstraint).isActive = true
+        bookimage.leftAnchor.constraint(equalTo: self.contentView.leftAnchor, constant: NumberManager.bookImageLeftConstraintInCell).isActive = true
+        bookimage.widthAnchor.constraint(equalToConstant: NumberManager.bookImageWidthConstraintInCell).isActive = true
+        bookimage.heightAnchor.constraint(equalToConstant: NumberManager.bookImageHeightConstraintInCell).isActive = true
         
         contentView.addSubview(title)
-        title.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: CGFloat(NumberManager.titleTopConstraint)).isActive = true
-        title.leftAnchor.constraint(equalTo: bookimage.rightAnchor, constant: CGFloat(NumberManager.titleLeftConstraint)).isActive = true
+        title.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: NumberManager.titleTopConstraint).isActive = true
+        title.leftAnchor.constraint(equalTo: bookimage.rightAnchor, constant: NumberManager.titleLeftConstraint).isActive = true
         
         contentView.addSubview(price)
-        price.topAnchor.constraint(equalTo: title.bottomAnchor, constant: CGFloat(NumberManager.priceTopConstraint)).isActive = true
-        price.leftAnchor.constraint(equalTo: bookimage.rightAnchor, constant: CGFloat(NumberManager.priceLeftConstraint)).isActive = true
+        price.topAnchor.constraint(equalTo: title.bottomAnchor, constant: NumberManager.priceTopConstraint).isActive = true
+        price.leftAnchor.constraint(equalTo: bookimage.rightAnchor, constant: NumberManager.priceLeftConstraint).isActive = true
         
         contentView.addSubview(bookDay)
-        bookDay.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: CGFloat(NumberManager.bookDayTopConstraint)).isActive = true
-        bookDay.leftAnchor.constraint(equalTo: price.rightAnchor, constant: CGFloat(NumberManager.bookDayLeftConstraint)).isActive = true
+        bookDay.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: NumberManager.bookDayTopConstraint).isActive = true
+        bookDay.leftAnchor.constraint(equalTo: price.rightAnchor, constant: NumberManager.bookDayLeftConstraint).isActive = true
         
         contentView.addSubview(editButton)
-        editButton.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: CGFloat(NumberManager.editButtonTopConstrain)).isActive = true
-        editButton.rightAnchor.constraint(equalTo: self.contentView.rightAnchor, constant: CGFloat(NumberManager.editButtonRightConstraint)).isActive = true
+        editButton.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: NumberManager.editButtonTopConstrain).isActive = true
+        editButton.rightAnchor.constraint(equalTo: self.contentView.rightAnchor, constant: NumberManager.editButtonRightConstraint).isActive = true
     }
 }
